Changed: Q3Support completely removed from project!
git-svn-id: file:///.data/svn_convert/svn@8659 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-playground/plugins/icq/icqmessage.cpp
@@ -349,7 +349,7 @@ Message *ICQClient::parseExtendedMessage(const QString &screen, ICQBuffer &packe
 	ICQBuffer b(info.size());
 	b.pack(info.c_str(), info.size());
 
-	log(L_DEBUG, "Extended message %s [%04X] %u", msgType.data(), msg_type, info.length());
+    log(L_DEBUG, "Extended message %s [%04X] %u", msgType.data(), msg_type, (unsigned int)info.length());
 
     int n = msgType.find("URL");
     if (n >= 0){
